book/examples/ai-coach/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
sqlalchemy==2.0.35
alembic==1.13.0
aiosqlite==0.20.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI / RAG
openai==1.50.0
llama-index==0.11.0
llama-index-llms-openai==0.2.0
llama-index-embeddings-openai==0.2.0
llama-index-vector-stores-chroma==0.2.0
llama-index-readers-file==0.2.0
chromadb==0.5.0

# Utils
httpx==0.27.0
apscheduler==3.10.4
loguru==0.7.2
